requirements.txt
# Core
MetaTrader5>=5.0.45
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
python-dateutil>=2.8.2

# Optional (для разработки)
# pytest>=7.0.0
# black>=23.0.0
# flake8>=6.0.0